% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_biphasic}
\alias{antibody_model_biphasic}
\title{Biphasic antibody boosting-waning model}
\usage{
antibody_model_biphasic(
  i,
  t1,
  ag,
  exposure_histories,
  antibody_states,
  kinetics_parameters,
  biomarker_map,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t1}{time}

\item{ag}{biomarker}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{antibody_states}{An array of antibody states across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{An object of all kinetics parameters for all exposures}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{}
}
\value{
A titer value is returned
}
\description{
Biphasic antibody boosting-waning model. This model assumes that for each exposure there is a set of long-term boost, long-term boost waning, short-term boost, and short-term boost waning parameters
}
