% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_biphasic_cpp}
\alias{antibody_model_biphasic_cpp}
\title{Biphasic antibody boosting-waning model Rcpp implementation}
\usage{
antibody_model_biphasic_cpp(
  i,
  t1,
  b,
  immune_histories,
  biomarker_states,
  kinetics_parameters,
  biomarker_map,
  ...
)
}
\arguments{
\item{i}{individual}

\item{t1}{time}

\item{b}{biomarker}

\item{immune_histories}{An array of immune histories across all individuals, time steps and exposure IDs}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all biomarkers}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{Additional arguments}
}
\value{
Biomarker quantity at the specified time point
}
\description{
Identical to \code{\link{antibody_model_biphasic}}, but implemented in Cpp
}
\examples{
tmp_pars <- list()
tmp_pars[[1]] <- draw_parameters_fixed_fx(1,1,1,NULL, NULL, example_model_pars_numeric)
antibody_model_biphasic_cpp(1,1,1,example_immune_histories_wide, example_biomarker_states_wide,tmp_pars, example_biomarker_map_numeric)
}
\seealso{
Other antibody_models: 
\code{\link{antibody_model_biphasic_cross_reactivity}()},
\code{\link{antibody_model_biphasic}()},
\code{\link{antibody_model_monophasic_cpp}()},
\code{\link{antibody_model_monophasic_cross_reactivity}()},
\code{\link{antibody_model_monophasic}()},
\code{\link{antibody_model_typhoid}()}
}
\concept{antibody_models}
