% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_monophasic}
\alias{antibody_model_monophasic}
\title{Monophasic antibody boosting-waning model}
\usage{
antibody_model_monophasic(
  i,
  t1,
  ag,
  exposure_histories,
  antibody_states,
  kinetics_parameters,
  antigen_map,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t1}{time}

\item{ag}{antigen}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{antibody_states}{An array of antibody states across all individuals, time steps and antigen IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all antigens}

\item{antigen_map}{A table specifying the relationship between exposure IDs and antigen IDs}

\item{...}{}
}
\value{
A titer value is returned
}
\description{
This monophasic antibody boosting-waning model model assumes that for each exposure there is a boost and boost waning parameter
}
