% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_monophasic_cross_reactivity}
\alias{antibody_model_monophasic_cross_reactivity}
\title{Monophasic antibody boosting-waning model with cross-reactive strains}
\usage{
antibody_model_monophasic_cross_reactivity(
  i,
  t1,
  b,
  immune_histories,
  biomarker_states,
  kinetics_parameters,
  biomarker_map,
  ...
)
}
\arguments{
\item{i}{individual}

\item{t1}{time}

\item{b}{biomarker}

\item{immune_histories}{An array of immune histories across all individuals, time steps and exposure IDs}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all biomarkers}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{Additional arguments}
}
\value{
A biomarker quantity is returned
}
\description{
Monophasic antibody boosting-waning model with cross-reactivity between strains. This monophasic antibody boosting-waning model assumes that for each exposure there is a boost and waning parameter describing antibody kinetics against the infecting strain (i.e., for exposure_id==biomarker_id). The model loops through each exposure type and reduces the amount of boosting as a function of cross-reactivity, which is determined by a proportion given in the \code{biomarker_map} data frame as the \code{value} variable.
}
\examples{
tmp_pars <- list()
## Set up simple model_pars table for this antibody model
library(dplyr)
model_pars_tmp <- example_model_pars_numeric \%>\% mutate(biomarker_id = 
exposure_id)
## Simulate one infection with exposure ID 1 at t=1
tmp_pars[[1]] <- draw_parameters_fixed_fx(1,1,1,NULL, NULL, model_pars_tmp)
 
## Set up a simple biomarker map for cross-reactivity
biomarker_map = tidyr::expand_grid(exposure_id=1:2, biomarker_id=1:2) \%>\% 
mutate(value = if_else(exposure_id==biomarker_id, 1, 0.5))
antibody_model_monophasic_cross_reactivity(1,1,1,
example_immune_histories_wide, example_biomarker_states_wide, 
tmp_pars, biomarker_map)
}
\seealso{
Other antibody_models: 
\code{\link{antibody_model_biphasic_cpp}()},
\code{\link{antibody_model_biphasic_cross_reactivity}()},
\code{\link{antibody_model_biphasic}()},
\code{\link{antibody_model_monophasic_cpp}()},
\code{\link{antibody_model_monophasic}()},
\code{\link{antibody_model_typhoid}()}
}
\concept{antibody_models}
