% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_typhoid}
\alias{antibody_model_typhoid}
\title{Power law antibody boosting and waning}
\usage{
antibody_model_typhoid(
  i,
  t,
  b,
  exposure_histories = NULL,
  biomarker_states = NULL,
  kinetics_parameters,
  biomarker_map = NULL,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{b}{biomarker}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all biomarkers}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{Additional arguments}
}
\value{
A biomarker_quantity is returned
}
\description{
Power law antibody boosting and waning model
}
\examples{
model_pars <- readr::read_csv("inst/extdata/model_pars_typhoid.csv")
tmp_pars <- list()
tmp_pars[[1]] <- draw_parameters_random_fx(1,1,1,1,NULL,NULL,model_pars)
tmp_exposure_history <- array(0,dim=c(1,11,2))
tmp_exposure_history[1,1,1] <- 1
antibody_model_typhoid(1,10, 1, tmp_exposure_history, NULL, tmp_pars,example_biomarker_map_numeric)
}
