% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{biomarker_protection}
\alias{biomarker_protection}
\title{Function to calculate the risk of infection at a given biomarker quantity}
\usage{
biomarker_protection(
  biomarker_quantity,
  biomarker_prot_midpoint,
  biomarker_prot_width
)
}
\arguments{
\item{biomarker_quantity}{Current biomarker quantity at the time of exposure}

\item{biomarker_prot_midpoint}{The biomarker quantity at which you are 50\% protected from infection}

\item{biomarker_prot_width}{Determines the shape of the curve}
}
\value{
The risk of infection given the current biomarker quantity
}
\description{
This function is used for the immunity models that incorporate biomarker mediated protection.
}
\examples{
biomarker_protection(50,25,.02)
}
