% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_titre.R
\name{calculate_titre}
\alias{calculate_titre}
\title{Calculates An Individual's Current Titre}
\usage{
calculate_titre(
  boosts_long_infections,
  boosts_short_infections,
  wanes_long_infections,
  wanes_short_infections,
  boosts_long_vacc,
  boosts_short_vacc,
  wanes_long_vacc,
  wanes_short_vacc,
  times,
  pathogen,
  i,
  cur_t
)
}
\arguments{
\item{boosts_long_infections}{A matrix containing all long-term boost parameters received from an infection at all possible infection times for each individual and pathogen}

\item{boosts_short_infections}{A matrix containing all short-term boost parameters received from an infection at all possible infection times for each individual and pathogen}

\item{wanes_long_infections}{A matrix containing all long-term boost waning parameters from an infection at all possible infection times for each individual and pathogen}

\item{wanes_short_infections}{A matrix containing all short-term boost waning parameters from an infection at all possible infection times for each individual and pathogen}

\item{boosts_long_vacc}{A matrix containing all long-term boost parameters received from vaccination at all possible vaccination times for each individual and pathogen}

\item{boosts_short_vacc}{A matrix containing all short-term boost parameters received from vaccination at all possible vaccination times for each individual and pathogen}

\item{wanes_long_vacc}{A matrix containing all long-term boost waning parameters from vaccination at all possible vaccination times for each individual and pathogen}

\item{wanes_short_vacc}{A matrix containing all short-term boost waning parameters from vaccination at all possible vaccination times for each individual and pathogen}

\item{times}{The total number of time steps for which you want to calculate an individual's titre}

\item{pathogen}{The pathogen(s) whose titre you want to measure; numeric entries only}

\item{i}{The individual}

\item{cur_t}{The current time step}
}
\value{
A titre level gets printed
}
\description{
Calculates An Individual's Current Titre
}
