% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{convert_indices_matrix_to_vector}
\alias{convert_indices_matrix_to_vector}
\title{Convert 2D indexing to 1D indexing}
\usage{
convert_indices_matrix_to_vector(i, j, n)
}
\arguments{
\item{i}{the row index}

\item{j}{the column index}

\item{n}{the number of rows}
}
\value{
the 1D indices corresponding to a collapsed matrix
}
\description{
R can treat matrices and arrays as 1D vectors, which can speed up indexing. This takes the row and column of the matrix and the number of rows (as R matrices are column-indexed) and returns the corresponding 1D index.
}
\examples{
x <- matrix(runif(10*5),nrow=10,ncol=5)
x[3,3]
x[convert_indices_matrix_to_vector(3,3,10)]
}
