% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_infection_exposure_parameters.R
\name{create_infection_exposure_parameters}
\alias{create_infection_exposure_parameters}
\title{Create Antibody Kinetics Parameters For Infection Boosts}
\usage{
create_infection_exposure_parameters(
  boosts_long_infections,
  boosts_short_infections,
  wanes_long_infections,
  wanes_short_infections,
  pathogen,
  i,
  t,
  kinetics_pars,
  current_titre
)
}
\arguments{
\item{boosts_long_infections}{Empty matrices for all possible infection times for each pathogen and individual to store infection long-term boost parameters}

\item{boosts_short_infections}{Empty matrices for all possible infection times for each pathogen and individual to store infection short-term boost parameters}

\item{wanes_long_infections}{Empty matrices for all possible infection times for each pathogen and individual to store infection long-term boost waning parameters}

\item{wanes_short_infections}{Empty matrices for all possible infection times for each pathogen and individual to store infection short-term boost waning parameters}

\item{pathogen}{The pathogen(s) whose infection parameters you want to generate; numeric entries only}

\item{i}{Individual}

\item{t}{The current time step}

\item{kinetics_pars}{The file with the antibody kinetics parameters}

\item{current_titre}{}
}
\value{
The 4 matrices are returned with parameters for all possible infection times for each individual and pathogen
}
\description{
Create Antibody Kinetics Parameters For Infection Boosts
}
