% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create Matrices To Store Simulation Information}
\usage{
create_matrix(N_pathogens, N, times, fill = NA)
}
\arguments{
\item{N_pathogens}{The number of pathogens in the simulation}

\item{N}{The number of individuals in the simulation}

\item{times}{The number of time steps in the simulation}

\item{fill}{Item to fill in matrix with; defaults to NA}
}
\value{
An empty matrix is returned with the necessary dimensions to store the simulaiton information
}
\description{
Create Matrices To Store Simulation Information
}
