% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters.R
\name{draw_parameters}
\alias{draw_parameters}
\title{Draw A Random Parameter From A Distribution}
\usage{
draw_parameters(par1, par2, distribution = "log-normal")
}
\arguments{
\item{par1}{The mean of the distribution}

\item{par2}{The standard deviation of the distribution}

\item{distribution}{The distribution to use; defaults to "log-normal"}
}
\value{
A parameter value gets printed
}
\description{
Draw A Random Parameter From A Distribution
}
\examples{
draw_parameters(5000,400,"log-normal")
draw_parameters(5000,400,"normal")
}
