% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_fixed_fx}
\alias{draw_parameters_fixed_fx}
\title{Draw parameters fixed effects}
\usage{
draw_parameters_fixed_fx(
  i,
  t,
  x,
  b,
  demography,
  biomarker_states,
  model_pars,
  ...
)
}
\arguments{
\item{i}{individual}

\item{t}{time}

\item{x}{exposure}

\item{b}{biomarker}

\item{demography}{demography information}

\item{biomarker_states}{an array of true biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{tibble of biomarker (antibody) kinetics parameters}

\item{...}{}
}
\value{
A tibble with the drawn parameters
}
\description{
Draws parameters directly from the distributions specified by \code{model_pars} for the antibody model with fixed effects. This function ensures that all individuals have the same parameters.
}
\examples{
draw_parameters_fixed_fx(1,1,1,1,demographu)
}
