% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_fixed_fx_biomarker_dep}
\alias{draw_parameters_fixed_fx_biomarker_dep}
\title{Draw Parameters Fixed Effects With Biomarker Quantity Dependent Boosting}
\usage{
draw_parameters_fixed_fx_biomarker_dep(
  i,
  t,
  x,
  b,
  demography,
  biomarker_states,
  model_pars,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{b}{biomarker}

\item{demography}{Demography information}

\item{biomarker_states}{An array of true biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of biomarker (antibody) kinetics parameters}

\item{...}{}
}
\value{
A tibble with the parameters drawn is returned
}
\description{
This function adds biomarker quantity ceiling effects to the previous draw_parameters_fixed_fx function. Here an individual’s realized biomarker boost is dependent on their biomarker quantity at the time of the exposure event.
}
