% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_random_fx}
\alias{draw_parameters_random_fx}
\title{Draw Parameters Random Effects}
\usage{
draw_parameters_random_fx(i, t, e, ag, demography, antibody_states, theta, ...)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{ag}{antigen}

\item{demography}{Demography information}

\item{antibody_states}{An array of true antibody titers for all individuals across all time steps and antigens}

\item{theta}{Tibble of antibody kinetics parameters}

\item{...}{}
}
\value{
A tibble with the parameters drawn is returned
}
\description{
This function draws parameters directly from theta for the antibody model with random effects. Parameters are drawn randomly from a distribution with mean and standard deviation specified within theta.
}
