% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_random_fx_titer_dep}
\alias{draw_parameters_random_fx_titer_dep}
\title{Draw Parameters Random Effects With Titer-Dependent Boosting}
\usage{
draw_parameters_random_fx_titer_dep(
  i,
  t,
  x,
  b,
  demography,
  antibody_states,
  model_pars,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{b}{biomarker}

\item{demography}{Demography information}

\item{antibody_states}{An array of true antibody titers for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of antibody kinetics parameters}

\item{...}{}
}
\value{
A tibble with the parameters drawn is returned
}
\description{
This function adds titer-ceiling effects to the previous draw_parameters_random_fx. Here an individual’s realized boost is dependent on their titer level at the time of the exposure event.
}
