% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_antigen_map}
\alias{example_antigen_map}
\title{Example antigen map}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{exposure_ID}{numeric values for each exposure type}
\item{antigen_ID}{numeric values for each antigen type present within each exposure type}
}
}
\usage{
data(example_antigen_map)
}
\description{
Example antigen map required as an input for runserosim. This antigen map is for a simulation with two different exposure events both containing the same antigen.
}
\seealso{
Other example_data: 
\code{\link{example_antibody_states}},
\code{\link{example_demography}},
\code{\link{example_exposure_histories}},
\code{\link{example_exposure_probabilities}},
\code{\link{example_foe_pars}},
\code{\link{example_model_pars}},
\code{\link{example_observed_antibody_states}}
}
\concept{example_data}
\keyword{datasets}
