% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_biomarker_states_wide}
\alias{example_biomarker_states_wide}
\title{Example biomarker states (wide)}
\format{
A 3D array with dimensions corresponding to 1) individual; 2) time; 3) biomarker, and values giving the latent biomarker quantity
}
\usage{
data(example_biomarker_states_wide)
}
\description{
Example biomarker states in wide format from \code{runserosim} output. This is identical to \code{example_biomarker_states}, but is converted into a 3D array giving the format expected by most antibody model functions. This example data set output contains biomarker quantities for individuals at all time steps for all biomarkers listed in \code{example_biomarker_map}
}
\seealso{
Other example_data: 
\code{\link{example_biomarker_map_numeric}},
\code{\link{example_biomarker_map}},
\code{\link{example_biomarker_states}},
\code{\link{example_demography}},
\code{\link{example_exposure_force}},
\code{\link{example_exposure_histories_wide}},
\code{\link{example_exposure_histories}},
\code{\link{example_exposure_probabilities}},
\code{\link{example_foe_pars}},
\code{\link{example_model_pars_biphasic}},
\code{\link{example_model_pars_numeric}},
\code{\link{example_model_pars_typhoid}},
\code{\link{example_model_pars}},
\code{\link{example_observed_biomarker_states}}
}
\concept{example_data}
\keyword{datasets}
