% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_demography}
\alias{example_demography}
\title{Example demography tibble}
\format{
A data frame with 12000 rows and 4 variables:
\describe{
\item{i}{string numeric values for each individual}
\item{times}{string numeric values for each time step in the simulation}
\item{birth}{numeric value indicating that individual's birth time}
\item{removal}{numeric value indicating that individual's removal time if any}
}
}
\usage{
data(example_demography)
}
\description{
Example of a demography tibble produced by generate_pop_demography and required as an input for runserosim
}
\seealso{
Other example_data: 
\code{\link{example_biomarker_map_numeric}},
\code{\link{example_biomarker_map}},
\code{\link{example_biomarker_states}},
\code{\link{example_exposure_force}},
\code{\link{example_exposure_histories}},
\code{\link{example_exposure_probabilities}},
\code{\link{example_foe_pars}},
\code{\link{example_model_pars_numeric}},
\code{\link{example_model_pars}},
\code{\link{example_observed_biomarker_states}}
}
\concept{example_data}
\keyword{datasets}
