% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_exposure_probabilities}
\alias{example_exposure_probabilities}
\title{Example exposure probability}
\format{
A data frame with 24000 rows and 4 variables:
\describe{
\item{i}{string numeric values for each individual}
\item{t}{string numeric values for each time step in the simulation}
\item{e}{numeric values for each exposure type}
\item{value}{numeric values indicating exposure probability}
}
}
\usage{
data(example_exposure_probabilities)
}
\description{
Example exposure probability runserosim output. This example data set output contains exposure probability for individuals at all time steps for each exposure ID
}
\seealso{
Other example_data: 
\code{\link{example_antibody_states}},
\code{\link{example_antigen_map}},
\code{\link{example_demography}},
\code{\link{example_exposure_histories}},
\code{\link{example_foe_pars}},
\code{\link{example_model_pars}},
\code{\link{example_observed_antibody_states}}
}
\concept{example_data}
\keyword{datasets}
