% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_immune_histories}
\alias{example_immune_histories}
\title{Example imune histories}
\format{
A data frame with 24000 rows and 4 variables:
\describe{
\item{i}{string numeric values for each individual}
\item{t}{string numeric values for each time step in the simulation}
\item{x}{numeric values for each exposure type}
\item{value}{binary value indicating if an exposure occured 1 or not 0}
}
}
\usage{
data(example_immune_histories)
}
\description{
Example immune history \code{runserosim} output. This example data set output contains immune history for individuals at all time steps for each exposure ID
}
\seealso{
Other example_data: 
\code{\link{example_biomarker_map_numeric}},
\code{\link{example_biomarker_map}},
\code{\link{example_biomarker_states_wide}},
\code{\link{example_biomarker_states}},
\code{\link{example_demography}},
\code{\link{example_exposure_force}},
\code{\link{example_exposure_probabilities}},
\code{\link{example_foe_pars}},
\code{\link{example_immune_histories_wide}},
\code{\link{example_model_pars_biphasic}},
\code{\link{example_model_pars_numeric}},
\code{\link{example_model_pars_typhoid}},
\code{\link{example_model_pars}},
\code{\link{example_observed_biomarker_states}}
}
\concept{example_data}
\keyword{datasets}
