% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model_pars}
\alias{example_model_pars}
\title{Example model parameters (model_pars)}
\format{
A data frame with 5 rows and 6 variables:
\describe{
\item{exposure_ID}{name of each exposure type present in example biomarker_map}
\item{biomarker_ID}{name of each biomarker present within each exposure type present in example biomarker_map}
\item{name}{names of model parameters}
\item{mean}{numeric values for the true paramter means}
\item{sd}{numeric values for the true paramter standard deviation}
\item{distribution}{distribution type from which the draw_paramaters function will be sampling a parameter }
}
}
\usage{
data(example_model_pars)
}
\description{
Example model parameters (model_pars). This model_pars input must be converted to the numeric version (example_model_pars_numeric) before being input into runserosim. This example model_pars is for a simulation two different exposure events both containing the same biomarker. model_pars argument is responsible for storing parameter information needed for the antibody model, observation model and immunity model.
In this example, model_pars has parameters for a monophasic antibody model and for an observational model with noise.
}
\seealso{
Other example_data: 
\code{\link{example_biomarker_map_numeric}},
\code{\link{example_biomarker_map}},
\code{\link{example_biomarker_states}},
\code{\link{example_demography}},
\code{\link{example_exposure_force}},
\code{\link{example_exposure_histories}},
\code{\link{example_exposure_probabilities}},
\code{\link{example_foe_pars}},
\code{\link{example_model_pars_numeric}},
\code{\link{example_observed_biomarker_states}}
}
\concept{example_data}
\keyword{datasets}
