% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_fixed}
\alias{exposure_model_fixed}
\title{Fixed probability of exposure}
\usage{
exposure_model_fixed(i, t, x, g, foe_pars, demography, ...)
}
\arguments{
\item{i}{individual}

\item{t}{time}

\item{x}{exposure}

\item{g}{group}

\item{foe_pars}{A 3D array providing the \emph{probability} of exposure for each exposure ID, group and time.}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{...}{}
}
\value{
A probability of exposure from the foe_pars array
}
\description{
Generic wrapper function to allow the probability of exposure for each time point to be drawn directly from the foe_pars array.
}
\examples{
times <- seq(0,365,by=1)
n_groups <- 1
n_exposures <- 1
foe_pars <- array(0.01, dim=c(n_groups,length(times),n_exposures))
exposure_model_fixed(1,1,1,1,foe_pars,NULL) 
}
