% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_simple_FOE}
\alias{exposure_model_simple_FOE}
\title{Exposure Model Simple- Force of Exposure (FOE)}
\usage{
exposure_model_simple_FOE(i, t, e, g, foe_pars, demography, ...)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{g}{group}

\item{foe_pars}{A 3D array providing the force of exposure for each exposure ID, group and time.}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{...}{}
}
\value{
A probability of exposure is returned
}
\description{
This is a simple exposure model where the probability of exposure depends on the force of exposure at that time for that group
}
