% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_simple_FOI}
\alias{exposure_model_simple_FOI}
\title{Exposure Model Simple- Force of Infection}
\usage{
exposure_model_simple_FOI(i, t, e, g, lambdas, demography, ...)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{g}{group}

\item{lambdas}{Force of infection array}

\item{demography}{Demography information}

\item{...}{}
}
\value{
A probability of exposure is returned
}
\description{
This is a simple exposure model where the probability of infection depends on the force of infection at that time for that group
}
