% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funcs.R
\name{get_infection_ratios}
\alias{get_infection_ratios}
\title{Get infection ratios}
\usage{
get_infection_ratios(chain, thin, burnin, adaptive_period, t1)
}
\description{
FluScape simulation specific function. Given an MCMC chain with burnin, thin and adaptive period, works out the proportion of infections that happened before t1 (ie. could have been observed). Does this for columns 2-4 of the chain.
}

