% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_all_successful}
\alias{immunity_model_all_successful}
\title{Immunity Model Where All Exposures Are Successful}
\usage{
immunity_model_all_successful(
  i,
  t,
  x,
  exposure_histories,
  biomarker_states,
  demography,
  biomarker_map,
  model_pars,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure events}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{model_pars}{A tibble of parameters needed for the immunity model}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
Probability of success is 1 so all exposure events are successful
}
