% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_ifxn_biomarker_prot}
\alias{immunity_model_ifxn_biomarker_prot}
\title{Immunity Model For Natural Infection Events With Biomarker Quantity Mediated Protection}
\usage{
immunity_model_ifxn_biomarker_prot(
  i,
  t,
  x,
  exposure_histories,
  biomarker_states,
  demography,
  biomarker_map,
  model_pars,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure events}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{biomarker_map}{A table specifying the relationship between exposure events and biomarker}

\item{model_pars}{A tibble of parameters needed for the immunity model}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
This immunity model should only be used for natural infection events. The probability of a successful exposure event is dependent on the individual’s biomarker quantity at the time of exposure. User specified biomakrer_prot_midpoint and biomarker_prot_width within model_pars is used to calculate biomarker-mediated protection.
}
