% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_ifxn_titer_prot}
\alias{immunity_model_ifxn_titer_prot}
\title{Immunity Model For Natural Infection Events With Titer-Mediated Protection}
\usage{
immunity_model_ifxn_titer_prot(
  i,
  t,
  e,
  exposure_histories,
  antibody_states,
  demography,
  antigen_map,
  theta,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{antibody_states}{True antibody titers for all individuals across all time steps and antigens}

\item{demography}{Demography information}

\item{antigen_map}{Object determining relationship between exposure IDs and antigens}

\item{theta}{Tibble including titer-mediated protection parameters}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
This immunity model should only be used if all exposures are natural infection events. The probability of successful exposure is dependent on the individual’s antibody titer at the time of exposure.
}
