% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_vacc_successful_ifxn}
\alias{immunity_model_vacc_successful_ifxn}
\title{Immunity Model For Vaccination Events and Successful Natural Infection Events}
\usage{
immunity_model_vacc_successful_ifxn(
  i,
  t,
  e,
  exposure_histories,
  antibody_states,
  demography,
  antigen_map,
  theta,
  max_vacc_events,
  vacc_exposures,
  vacc_age,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{antibody_states}{True antibody titers for all individuals across all time steps and antigens}

\item{demography}{Demography information}

\item{antigen_map}{Object determining relationship between exposure IDs and antigens}

\item{theta}{Tibble including titer-mediated protection parameters}

\item{max_vacc_events}{A vector of the maximum number of vaccination events possible for each exposure type; If an exposure type is not a vaccination event then input NA}

\item{vacc_exposures}{A vector of exposure IDs (e) which represents vaccination events}

\item{vacc_age}{A vector of the minimum age at which an individual is eligible for vaccination for each exposure type; If an exposure type is not a vaccination event then input NA}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
This immunity model should only be used with vaccine and natural infection. The probability of successful exposure(vaccination event) depends on the number of vaccines an individual has received prior to time t. If the individual is under the maximum vaccinations allotted then the probability of successful exposure is 1. The probability of a successful natural infection is 1.
}
