% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kinetics_parameters.R
\name{load_kinetics_parameters}
\alias{load_kinetics_parameters}
\title{Import And Define Antibody Kinetics Parameters}
\usage{
load_kinetics_parameters(
  boost_infection_long_mean_1 = 0,
  boost_infection_long_var_1 = 0,
  wane_infection_long_mean_1 = 0,
  wane_infection_long_var_1 = 0,
  boost_infection_short_mean_1 = 0,
  boost_infection_short_var_1 = 0,
  wane_infection_short_mean_1 = 0,
  wane_infection_short_var_1 = 0,
  boost_vacc_long_mean_1 = 0,
  boost_vacc_long_var_1 = 0,
  wane_vacc_long_mean_1 = 0,
  wane_vacc_long_var_1 = 0,
  boost_vacc_short_mean_1 = 0,
  boost_vacc_short_var_1 = 0,
  wane_vacc_short_mean_1 = 0,
  wane_vacc_short_var_1 = 0,
  titre_ceiling_gradient_1 = 0,
  titre_ceiling_threshold_1 = 0,
  boost_infection_long_mean_2 = 0,
  boost_infection_long_var_2 = 0,
  wane_infection_long_mean_2 = 0,
  wane_infection_long_var_2 = 0,
  boost_infection_short_mean_2 = 0,
  boost_infection_short_var_2 = 0,
  wane_infection_short_mean_2 = 0,
  wane_infection_short_var_2 = 0,
  boost_vacc_long_mean_2 = 0,
  boost_vacc_long_var_2 = 0,
  wane_vacc_long_mean_2 = 0,
  wane_vacc_long_var_2 = 0,
  boost_vacc_short_mean_2 = 0,
  boost_vacc_short_var_2 = 0,
  wane_vacc_short_mean_2 = 0,
  wane_vacc_short_var_2 = 0,
  titre_ceiling_gradient_2 = 0,
  titre_ceiling_threshold_2 = 0
)
}
\arguments{
\item{boost_infection_long_mean_1}{The mean parameter for infection induced long-term boost for pathogen 1; defaults to 0}

\item{boost_infection_long_var_1}{The variance parameter for infection induced long-term boost for pathogen 1; defaults to 0}

\item{wane_infection_long_mean_1}{The mean parameter for infection induced long-term boost waning for pathogen 1; defaults to 0}

\item{wane_infection_long_var_1}{The variance parameter for infection induced long-term boost waning for pathogen 1; defaults to 0}

\item{boost_infection_short_mean_1}{The mean parameter for infection induced short-term boost for pathogen 1; defaults to 0}

\item{boost_infection_short_var_1}{The variance parameter for infection induced short-term boost for pathogen 1; defaults to 0}

\item{wane_infection_short_mean_1}{The mean parameter for infection induced short-term boost waning for pathogen 1; defaults to 0}

\item{wane_infection_short_var_1}{The variance parameter for infection induced short-term boost waning for pathogen 1; defaults to 0}

\item{boost_vacc_long_mean_1}{The mean parameter for vaccination induced long-term boost for pathogen 1; defaults to 0}

\item{boost_vacc_long_var_1}{The variance parameter for vaccination induced long-term boost for pathogen 1; defaults to 0}

\item{wane_vacc_long_mean_1}{The mean parameter for vaccination induced long-term boost waning for pathogen 1; defaults to 0}

\item{wane_vacc_long_var_1}{The variance parameter for vaccination induced long-term boost waning for pathogen 1; defaults to 0}

\item{boost_vacc_short_mean_1}{The mean parameter for vaccination induced short-term boost for pathogen 1; defaults to 0}

\item{boost_vacc_short_var_1}{The variance parameter for vaccination induced short-term boost for pathogen 1; defaults to 0}

\item{wane_vacc_short_mean_1}{The mean parameter for vaccination induced short-term boost waning for pathogen 1; defaults to 0}

\item{wane_vacc_short_var_1}{The variance parameter for vaccination induced short-term boost waning for pathogen 1; defaults to 0}

\item{titre_ceiling_gradient_1}{(1-A)/B; Where A is the proportion of the full boost received at or above the titre_ceiling_threshold (B) for pathogen 1; defaults to 0}

\item{titre_ceiling_threshold_1}{The maximum titre level an individual can have before their secondary boost is limited in size for pathogen 1; defaults to 0}

\item{boost_infection_long_mean_2}{The mean parameter for infection induced long-term boost for pathogen 2; defaults to 0}

\item{boost_infection_long_var_2}{The variance parameter for infection induced long-term boost for pathogen 2; defaults to 0}

\item{wane_infection_long_mean_2}{The mean parameter for infection induced long-term boost waning for pathogen 2; defaults to 0}

\item{wane_infection_long_var_2}{The variance parameter for infection induced long-term boost waning for pathogen 2; defaults to 0}

\item{boost_infection_short_mean_2}{The mean parameter for infection induced short-term boost for pathogen 2; defaults to 0}

\item{boost_infection_short_var_2}{The variance parameter for infection induced short-term boost for pathogen 2; defaults to 0}

\item{wane_infection_short_mean_2}{The mean parameter for infection induced short-term boost waning for pathogen 2; defaults to 0}

\item{wane_infection_short_var_2}{The variance parameter for infection induced short-term boost waning for pathogen 2; defaults to 0}

\item{boost_vacc_long_mean_2}{The mean parameter for vaccination induced long-term boost for pathogen 2; defaults to 0}

\item{boost_vacc_long_var_2}{The variance parameter for vaccination induced long-term boost for pathogen 2; defaults to 0}

\item{wane_vacc_long_mean_2}{The mean parameter for vaccination induced long-term boost waning for pathogen 2; defaults to 0}

\item{wane_vacc_long_var_2}{The variance parameter for vaccination induced long-term boost waning for pathogen 2; defaults to 0}

\item{boost_vacc_short_mean_2}{The mean parameter for vaccination induced short-term boost for pathogen 2; defaults to 0}

\item{boost_vacc_short_var_2}{The variance parameter for vaccination induced short-term boost for pathogen 2; defaults to 0}

\item{wane_vacc_short_mean_2}{The mean parameter for vaccination induced short-term boost waning for pathogen 2; defaults to 0}

\item{wane_vacc_short_var_2}{The variance parameter for vaccination induced short-term boost waning for pathogen 2; defaults to 0}

\item{titre_ceiling_gradient_2}{(1-A)/B; Where A is the proportion of the full boost received at or above the titre_ceiling_threshold (B) for pathogen 2; defaults to 0}

\item{titre_ceiling_threshold_2}{The maximum titre level an individual can have before their secondary boost is limited in size for pathogen 2; defaults to 0}
}
\value{
A data frame with kinetics parameters for each pathogen is returned
}
\description{
Import And Define Antibody Kinetics Parameters
}
