% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funcs.R
\name{multiple_strains}
\alias{multiple_strains}
\title{Boost/wane model}
\usage{
multiple_strains(tis, y0, params, times)
}
\arguments{
\item{tis}{vector of times of infection eg. c(0, 45, 70)}

\item{y0}{vector of baseline HAI values for each infecting strain}

\item{params}{list of parameters for the model. This should be params[[1]] = mu_pars (matrix); params[[2]] = tp_pars (matrix); params[[3]] = m_pars (matrix); params[[4]] = STOCHASTIC (bool).}

\item{times}{vector of times in days.}
}
\value{
matrix of times and longitundinal simulation for each infecting strain
}
\description{
Data generating function for a multiple strain, boosting and waning serological model.
}
\examples{
test
Create matrix with a column for each infection and a column for times
Sort infection times into ascending order and store indices of sorted order. Add this to list of process parameters if needed
Extract parameters from list
For each strain, calculate the single trajectory
If stochastic boosting required, generate mu from poisson distribution
}
\seealso{
\code{\link{single_strain}}
}

