% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_to_lognormal_mean.R
\name{normal_to_lognormal_mean}
\alias{normal_to_lognormal_mean}
\title{Convert Normal Distribution Mean to Log-Normal Distribution Mean}
\usage{
normal_to_lognormal_mean(normmean, normsd)
}
\arguments{
\item{normmean}{The normal distribution mean}

\item{normsd}{The normal distribution standard deviation}
}
\value{
The log-normal distribution mean is printed
}
\description{
Convert Normal Distribution Mean to Log-Normal Distribution Mean
}
\examples{
normal_to_lognormal_mean(5000,400)
}
