% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{normal_to_lognormal_sd}
\alias{normal_to_lognormal_sd}
\title{Convert normal distributions to log-normal distributions and return standard deviation}
\usage{
normal_to_lognormal_sd(normmean, normsd)
}
\arguments{
\item{normmean}{Mean of the normal distribution}

\item{normsd}{Standard deviation of the normal distribution}
}
\value{
Standard deviation of the log-normal distribution
}
\description{
Convert normal distributions to log-normal distributions and return standard deviation
}
\examples{
normal_to_lognormal_sd(50,10)
}
