% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{observation_model_continuous_bounded}
\alias{observation_model_continuous_bounded}
\title{Observation Model For Continuous Assays With Detection Limits And No Added Noise}
\usage{
observation_model_continuous_bounded(biomarker_states, model_pars, bounds, ...)
}
\arguments{
\item{biomarker_states}{True biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of observation model parameters}

\item{bounds}{A tibble containing the assay lower bound and upper bound for all biomarkers; column names=biomarker_id, name, and value where name is either "lower_bound" or "upper_bound"}

\item{...}{}
}
\value{
biomarker_states is returned with a new column for observed biomarker quantities
}
\description{
This observation model observes the latent biomarker quantities given a continuous assay with user-specified lower and upper limits and no added noise.
}
