% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{observation_model_continuous_bounded_no_noise}
\alias{observation_model_continuous_bounded_no_noise}
\title{Observation Model For Continuous Assays With Detection Limits And No Added Noise}
\usage{
observation_model_continuous_bounded_no_noise(
  antibody_states,
  theta,
  demography,
  boundary,
  ...
)
}
\arguments{
\item{antibody_states}{True antibody titers for all individuals across all time steps and antigens}

\item{theta}{Tibble of observation model parameters}

\item{demography}{Demography information}

\item{boundary}{A list containing the lower bound and upper bound of the assay}

\item{...}{}
}
\value{
antibody_states is returned with a new column for observed titers
}
\description{
This observation model observes the latent titer values given a continuous assay with user-specified lower and upper limits and no added noise.
}
