% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_infection.R
\name{p_infection}
\alias{p_infection}
\title{Generate The Probability Of Infection At A Given Titre}
\usage{
p_infection(phi, titre, alpha1, beta1)
}
\arguments{
\item{phi}{The probability of infection}

\item{titre}{An individual's current titre level}

\item{alpha1}{The titre value at which the probability of infection is 0.5}

\item{beta1}{This value dictates the shape of the curve; smaller values like 0.001 will produce smoother curves}
}
\value{
The probability of infection is returned
}
\description{
Generate The Probability Of Infection At A Given Titre
}
