% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_infection_probs.R
\name{plot_all_infection_probs}
\alias{plot_all_infection_probs}
\title{Plot Infection Probabilities Across Time For All Individuals And Pathogens}
\usage{
plot_all_infection_probs(all_infection_probs)
}
\arguments{
\item{all_infection_probs}{The reshaped data set containing infection probability for individuals at all time steps for each pathogen}
}
\value{
A plot of infection probabilities across time for all individuals and pathogens is returned
}
\description{
Plot Infection Probabilities Across Time For All Individuals And Pathogens
}
