% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_biomarker_dependent_boosting}
\alias{plot_biomarker_dependent_boosting}
\title{Generate a plot displaying proportion of full boost received at each starting biomarker quantity}
\usage{
plot_biomarker_dependent_boosting(
  start,
  end,
  by,
  biomarker_ceiling_threshold,
  biomarker_ceiling_gradient
)
}
\arguments{
\item{start}{Lower bound of the x axis}

\item{end}{Upper bound of the x axis}

\item{by}{Increments at which your x axis will be plotted}

\item{biomarker_ceiling_threshold}{The maximum biomarker level an individual can have before their boost is limited in size}

\item{biomarker_ceiling_gradient}{(1-A)/B; Where A is the proportion of the full boost received at or above the biomarker_ceiling_threshold (B)}
}
\value{
A plot displaying biomarker quantity dependent boosting is returned
}
\description{
Generate a plot displaying proportion of full boost received at each starting biomarker quantity
}
\examples{
plot_biomarker_dependent_boosting(0,10,1,4,.14)
}
