% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_biomarker_mediated_protection}
\alias{plot_biomarker_mediated_protection}
\title{Plot Biomarker Mediated Protection Graphs For Each Pathogen}
\usage{
plot_biomarker_mediated_protection(
  biomarker_range,
  biomarker_prot_midpoint,
  biomarker_prot_width
)
}
\arguments{
\item{biomarker_range}{The range of possible biomarker quantities an individual can have at exposure}

\item{biomarker_prot_midpoint}{The biomarker quantity at which you are 50\% protected from infection}

\item{biomarker_prot_width}{Determines the shape of the curve}
}
\value{
A plot of the probability of infection given an individual's biomarker quantity at exposure
}
\description{
Plot Biomarker Mediated Protection Graphs For Each Pathogen
}
\examples{
plot_biomarker_mediated_protection(0:10,5,0.9)
}
