% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_biomarker_quantity}
\alias{plot_biomarker_quantity}
\title{Plot Biomarker Quantities Across Time For All Individuals And Biomarkers}
\usage{
plot_biomarker_quantity(biomarker_states)
}
\arguments{
\item{biomarker_states}{The reshaped data set containing biomarker quantity for individuals at all time steps for each biomarker}
}
\value{
A plot of biomarker quantities across all time steps for all individuals and biomarkers is returned
}
\description{
Plot Biomarker Quantities Across Time For All Individuals And Biomarkers
}
\examples{
plot_biomarker_quantity(example_biomarker_states)
}
