% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_exposure_prob}
\alias{plot_exposure_prob}
\title{Plot probability of a successful exposure event for all individuals and exposure events}
\usage{
plot_exposure_prob(exposure_probabilities_long)
}
\arguments{
\item{exposure_probabilities_long}{The reshaped data set containing the probability of a successful exposure event for individuals at all time steps for each exposure event}
}
\value{
A plot of the probability of a successful exposure event across time for all individuals and exposure events is returned
}
\description{
Plot probability of a successful exposure event for all individuals and exposure events
}
\examples{
plot_exposure_prob(example_exposure_probabilities)
}
