% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_immune_histories}
\alias{plot_immune_histories}
\title{Plot individual immune histories}
\usage{
plot_immune_histories(immune_histories)
}
\arguments{
\item{immune_histories}{The reshaped data set containing immune history for individuals at all time steps for each exposure event}
}
\value{
A plot of individual immune histories across time for all individuals and exposure events is returned
}
\description{
Plot individual immune histories
}
\examples{
plot_immune_histories(example_immune_histories)
}
