% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_obs_biomarkers_one_sample}
\alias{plot_obs_biomarkers_one_sample}
\title{Plot Observed Biomarker Quantities For One Observation Time}
\usage{
plot_obs_biomarkers_one_sample(observed_biomarker_states)
}
\arguments{
\item{observed_biomarker_states}{The reshaped data set containing observed biomarker quantities for individuals at all time steps for each biomarker}
}
\value{
A plot of observed biomarker quantities for all individuals and biomarkers is returned
}
\description{
This function should be used when there was only one time step in which biomarker quantities were observed
}
\examples{
plot_obs_biomarkers_one_sample(example_observed_biomarker_states)
}
