% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_obs_titers_one_sample}
\alias{plot_obs_titers_one_sample}
\title{Plot Observed Titers For One Observation Time}
\usage{
plot_obs_titers_one_sample(observed_titers)
}
\arguments{
\item{observed_titers}{The reshaped data set containing observed antibody titers for individuals at all time steps for each antigen}
}
\value{
A plot of observed titers for all individuals and antigens is returned
}
\description{
This function should be used when there was only one time step in which titers were observed
}
