% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_obs_titers_paired_sample}
\alias{plot_obs_titers_paired_sample}
\title{Plot Observed Titers For Multiple Observation Times and Paired Samples}
\usage{
plot_obs_titers_paired_sample(observed_titers)
}
\arguments{
\item{observed_titers}{The reshaped data set containing observed antibody titers for individuals at all time steps for each biomarker}
}
\value{
A plot of observed titers for all individuals and biomarkers is returned
}
\description{
This function should be used when there were multiple time step in which titers were observed
}
