% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_serosurvey_distribution.R
\name{plot_serosurvey_distribution}
\alias{plot_serosurvey_distribution}
\title{Plot Serosurvey Titre Distribution}
\usage{
plot_serosurvey_distribution(
  titresobs,
  titre = "observed",
  seronegative = "keep",
  bin,
  color = "pathogen"
)
}
\arguments{
\item{titresobs}{The titre data set for all individuals at the observation time}

\item{titre}{Indicate whether you want to plot "true" or "observed" titres; defaults to "observed"}

\item{seronegative}{Indicate whether you want to keep or remove seronegative individuals; defaults to "keep"}

\item{bin}{The size of the bins for the histogram}

\item{color}{The variable to label the points either by "events" or "pathogen"; defaults to "pathogen"}
}
\value{
A histogram is returned
}
\description{
Plot Serosurvey Titre Distribution
}
