% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_titer_dependent_boosting}
\alias{plot_titer_dependent_boosting}
\title{Generate A Plot Displaying Proportion Of Full Boost Received At Each Starting Titer}
\usage{
plot_titer_dependent_boosting(
  start,
  end,
  by,
  titer_ceiling_threshold,
  titer_ceiling_gradient
)
}
\arguments{
\item{start}{Lower bound of the x axis}

\item{end}{Upper bound of the x axis}

\item{by}{Increments at which your x axis will be plotted}

\item{titer_ceiling_threshold}{The maximum titer level an individual can have before their boost is limited in size}

\item{titer_ceiling_gradient}{(1-A)/B; Where A is the proportion of the full boost received at or above the titer_ceiling_threshold (B)}
}
\value{
A plot displaying titer dependent boosting is returned
}
\description{
Generate A Plot Displaying Proportion Of Full Boost Received At Each Starting Titer
}
