% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_titer_mediated_protection}
\alias{plot_titer_mediated_protection}
\title{Plot Titer Mediated Protection Graphs For Each Pathogen}
\usage{
plot_titer_mediated_protection(
  titer_range,
  titer_prot_midpoint,
  titer_prot_width
)
}
\arguments{
\item{titer_range}{The range of possible titers an individual can have at exposure}

\item{titer_prot_midpoint}{The titer value at which you are 50\% protected from infection}

\item{titer_prot_width}{Determines the chape of the curve}
}
\value{
A plot of the probability of infection given an individual's titer at exposure
}
\description{
Plot Titer Mediated Protection Graphs For Each Pathogen
}
\examples{
plot_titer_mediated_protection(0:10,5,0.9)
}
