% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_titers}
\alias{plot_titers}
\title{Plot Titers Across Time For All Individuals And Antigens}
\usage{
plot_titers(titers)
}
\arguments{
\item{titers}{The reshaped data set containing antibody titer for individuals at all time steps for each antigen ID}
}
\value{
A plot of titers across all time steps for all individuals and antigens is returned
}
\description{
Plot Titers Across Time For All Individuals And Antigens
}
\examples{
plot_titers(example_antibody_states)
}
