% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_titre_dependent_boosting.R
\name{plot_titre_dependent_boosting}
\alias{plot_titre_dependent_boosting}
\title{Generate A Plot Displaying Proportion Of Boost Received At Each Titre}
\usage{
plot_titre_dependent_boosting(start, end, by, kinetics_pars, pathogen)
}
\arguments{
\item{start}{This value is the lower bound of the x axis}

\item{end}{This value is the upper bound of the x axis}

\item{by}{This value indicates the increments at which your x axis will be plotted}

\item{kinetics_pars}{This is your antibody kinetics data frame}

\item{pathogen}{This is the pathogen you are interested in plotting; 1 or 2}
}
\value{
A plot displaying titre dependent boosting is returned
}
\description{
Generate A Plot Displaying Proportion Of Boost Received At Each Titre
}
