% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{reformat_antigen_map}
\alias{reformat_antigen_map}
\title{Reformat antigen_map to numeric exposure and antigen format needed for runserosim}
\usage{
reformat_antigen_map(antigen_map)
}
\arguments{
\item{antigen_map}{A table specifying the relationship between exposure IDs and antigen IDs}
}
\value{
antigen_map is returned with unique numeric inputs for exposure and antigen IDs
}
\description{
#' @description This function will reformat the antigen_map argument so that exposure_ID and antigen_ID are both numeric
}
