% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{reformat_biomarker_map}
\alias{reformat_biomarker_map}
\title{Reformat biomarker_map to numeric exposure and biomarker format needed for runserosim}
\usage{
reformat_biomarker_map(biomarker_map)
}
\arguments{
\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}
}
\value{
biomarker_map is returned with unique numeric inputs for exposure and biomarker IDs
}
\description{
This function will reformat the biomarker_map argument so that exposure_ID and biomarker_ID are both numeric
}
