% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{reformat_model_pars}
\alias{reformat_model_pars}
\title{Reformat model_pars to numeric exposure and antigen format needed for runserosim}
\usage{
reformat_model_pars(antigen_map, model_pars)
}
\arguments{
\item{antigen_map}{A table specifying the relationship between exposure IDs and antigen IDs}

\item{model_pars}{A tibble of parameters needed for the antibody kinetics model, immunity model, observation model and the draw_parameters function}
}
\value{
antigen_map is returned with unique numeric inputs for exposure and antigen IDs
}
\description{
This function will reformat the model_pars argument so that exposure_ID and antigen_ID are both numeric
}
