% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_infection_histories.R
\name{reshape_infection_histories}
\alias{reshape_infection_histories}
\title{Reshape Infection Histories Across Time For All Individuals And Pathogens}
\usage{
reshape_infection_histories(infection_histories, N_pathogens)
}
\arguments{
\item{infection_histories}{A matrix containing the infection histories for each individual at each time step}

\item{N_pathogens}{The number of pathogens in the simulation}
}
\value{
An infection history data set is returned
}
\description{
Reshape Infection Histories Across Time For All Individuals And Pathogens
}
