% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_titres.R
\name{reshape_titres}
\alias{reshape_titres}
\title{Reshape Titres Matrix}
\usage{
reshape_titres(titres, N_pathogens)
}
\arguments{
\item{titres}{A matrix containing the antibody titres for each individual at each time step}

\item{N_pathogens}{The number of pathogens in the simulation}
}
\value{
A titre data set is returned
}
\description{
Reshape Titres Matrix
}
