% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_vaccine_histories.R
\name{reshape_vaccine_histories}
\alias{reshape_vaccine_histories}
\title{Reshape Vaccine Histories Across Time For All Individuals And Pathogens}
\usage{
reshape_vaccine_histories(vaccine_histories)
}
\arguments{
\item{vaccine_histories}{A matrix containing the vaccination histories for each individual at each time step}
}
\value{
A vaccine history data set is returned
}
\description{
Reshape Vaccine Histories Across Time For All Individuals And Pathogens
}
