% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run Simulation To Generate Infection And Vaccination Events}
\usage{
run_simulation(
  birth_times,
  N,
  times,
  N_pathogens,
  kinetics_pars,
  vaccine_histories,
  pathogen,
  i,
  t
)
}
\arguments{
\item{birth_times}{Each individual's time of birth}

\item{N}{The number of individuals in the simulation}

\item{times}{The number of time steps in the simulation}

\item{N_pathogens}{The number of pathogens in the simulation}

\item{kinetics_pars}{A file with the antibody kinetics parameters}

\item{vaccine_histories}{A matrix with individual's vaccine history at each time step}

\item{pathogen}{Pathogen}

\item{i}{Individual}

\item{t}{Time step}
}
\value{
13 completed matrices are returned
}
\description{
Run Simulation To Generate Infection And Vaccination Events
}
