% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosim.R
\name{runserosim}
\alias{runserosim}
\title{Main simulation function for serosim}
\usage{
runserosim(
  simulation_settings,
  demography = NULL,
  observation_times = NULL,
  lambdas,
  antigen_map,
  theta,
  exposure_model,
  immunity_model,
  antibody_model,
  observation_model,
  draw_parameters,
  exposure_histories_fixed = NULL,
  ...
)
}
\arguments{
\item{exposure_histories_fixed}{}
}
\value{
a list containing the following elements.
}
\description{
Simulates a serological survey using custom inputs. The user can specify multiple inputs controlling population demography, simulation timeframe, observation times for each individual, force of infection, and various model functions describing the link between infections and observed antibody titers.
}
