% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{simulate_birth_times}
\alias{simulate_birth_times}
\title{Simulate Random Birth Times}
\usage{
simulate_birth_times(N, times, age_min = 0)
}
\arguments{
\item{N}{The number of individuals in the simulation}

\item{times}{A vector of each time step in the simulation}

\item{age_min}{A number indicating the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the second to last time step}
}
\value{
A vector of simulated birth times for each individual is returned
}
\description{
This function simulates random birth times for a specified number of individuals from a provided vector
}
\examples{
## Simulate random birth times for 500 individuals over 100 time steps and ensures that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
}
\seealso{
Other demography: 
\code{\link{generate_pop_demography}()},
\code{\link{simulate_removal_times}()}
}
\concept{demography}
