% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{update}
\alias{update}
\title{runserosim function update message}
\usage{
update(VERBOSE, i)
}
\arguments{
\item{VERBOSE}{The interval of individuals in which messages will be printed at}

\item{i}{The individual that the simulation is currently on}
}
\value{
A printed statement indicating an individual number is returned
}
\description{
This function is used within runserosim to produce messages to inform the user which individual the simulation is currently running
}
\examples{
## update(10, 100)
}
