% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_biphasic}
\alias{antibody_model_biphasic}
\title{Biphasic antibody boosting-waning model}
\usage{
antibody_model_biphasic(
  i,
  t1,
  b,
  exposure_histories,
  biomarker_states,
  kinetics_parameters,
  biomarker_map,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t1}{time}

\item{b}{biomarker}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all biomarkers}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{Additional arguments}
}
\value{
A biomarker quantity is returned
}
\description{
Biphasic antibody boosting-waning model. This model assumes that for each exposure there is a set of long-term boost, long-term boost waning, short-term boost, and short-term boost waning parameters
}
\examples{
model_pars <- reformat_biomarker_map(example_model_pars_biphasic)
tmp_pars <- list()
tmp_pars[[1]] <- draw_parameters_fixed_fx_biomarker_dep(1,1,1,1,NULL, NULL, model_pars)
antibody_model_biphasic(1,1,1,example_exposure_histories, example_biomarker_states, tmp_pars, example_biomarker_map_numeric)
}
