% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_monophasic}
\alias{antibody_model_monophasic}
\title{Monophasic antibody boosting-waning model}
\usage{
antibody_model_monophasic(
  i,
  t1,
  b,
  exposure_histories,
  biomarker_states,
  kinetics_parameters,
  biomarker_map,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t1}{time}

\item{b}{biomarker}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all biomarkers}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{Additional arguments}
}
\value{
A biomarker quantity is returned
}
\description{
This monophasic antibody boosting-waning model model assumes that for each exposure there is a boost and boost waning parameter
}
\examples{
tmp_pars <- list()
tmp_pars[[1]] <- draw_parameters_fixed_fx(1,1,1,1,NULL, NULL, example_model_pars_numeric)
antibody_model_monophasic(1,1,1,example_exposure_histories, example_biomarker_states, tmp_pars, example_biomarker_map_numeric)
}
