% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_models.R
\name{antibody_model_typhoid}
\alias{antibody_model_typhoid}
\title{Power law antibody boosting and waning}
\usage{
antibody_model_typhoid(
  i,
  t,
  b,
  exposure_histories = NULL,
  biomarker_states = NULL,
  kinetics_parameters,
  biomarker_map = NULL,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{b}{biomarker}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{biomarker_states}{An array of biomarker states (biomarker quantities) across all individuals, time steps and biomarker IDs}

\item{kinetics_parameters}{A tibble of parameters needed for the antibody kinetics model for all biomarkers}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{...}{}
}
\value{
A biomarker_quantity is returned
}
\description{

}
