% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_fixed_fx}
\alias{draw_parameters_fixed_fx}
\title{Draw Parameters Fixed Effects}
\usage{
draw_parameters_fixed_fx(
  i,
  t,
  e,
  ag,
  demography,
  antibody_states,
  model_pars,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{ag}{biomarker}

\item{demography}{Demography information}

\item{antibody_states}{An array of true antibody titers for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of antibody kinetics parameters}

\item{...}{}
}
\value{
A tibble with the parameters drawn is returned
}
\description{
This function draws parameters directly from model_pars for the antibody model with fixed effects. This function ensures that all individuals have the same parameters.
}
