% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_random_fx}
\alias{draw_parameters_random_fx}
\title{Draw parameters with random effects}
\usage{
draw_parameters_random_fx(
  i,
  t,
  x,
  b,
  demography,
  biomarker_states,
  model_pars,
  ...
)
}
\arguments{
\item{i}{individual}

\item{t}{time}

\item{x}{exposure}

\item{b}{biomarker}

\item{demography}{demography information}

\item{biomarker_states}{an array of true biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{tibble of biomarker (antibody) kinetics parameters with variables: 1) exposure_id: numeric exposure ID; 2) biomarker_id: numeric biomarker ID; 3) name: the character name of the parameter; 4) mean: numeric mean of this parameter distribution; 5) sd: the numeric standard deviation of the parameter distribution; 6) distribution: character description of the parameter distribution type (e.g., log-normal, normal)}

\item{...}{Additional arguments}
}
\value{
A tibble with the simulated parameters for this exposure event
}
\description{
Draws parameters directly from \code{model_pars} for the antibody model with random effects. Parameters are drawn randomly from a distribution with mean and standard deviation specified within \code{model_pars}.
}
\examples{
draw_parameters_random_fx(1,1,1,1,example_demography, example_biomarker_states, example_model_pars_numeric)
}
