% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_parameters_options.R
\name{draw_parameters_random_fx}
\alias{draw_parameters_random_fx}
\title{Draw Parameters Random Effects}
\usage{
draw_parameters_random_fx(
  i,
  t,
  x,
  b,
  demography,
  biomarker_states,
  model_pars,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{b}{biomarker}

\item{demography}{Demography information}

\item{biomarker_states}{An array of true biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of biomarker (antibody) kinetics parameters}

\item{...}{}
}
\value{
A tibble with the parameters drawn is returned
}
\description{
This function draws parameters directly from model_pars for the antibody model with random effects. Parameters are drawn randomly from a distribution with mean and standard deviation specified within model_pars.
}
