% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_foe_pars}
\alias{example_foe_pars}
\title{Example force of exposure parameters (\code{foe_pars})}
\format{
A 3 dimensional array containing the force of exposure at each time step for each exposure ID:
\describe{
\item{dimension 1}{group dimension is 1  since no groups were specified in example_demography. Since there are no specified groups, all individuals are under the same force of exposure parameter}
\item{dimenstion 2}{time step dimension is 120 since there are 120 time steps in the simulation}
\item{dimenstion 3}{exposure ID dimension for 2 exposure types}
}
}
\usage{
data(example_foe_pars)
}
\description{
Example force of exposure parameters required (\code{foe_pars}) as an input for \code{runserosim}. \code{foe_pars} is a 3D array providing the force of exposure for each exposure ID, group and time. This \code{foe_pars} argument is for a simulation with 120 time steps and two different exposure events (dimension 3)
}
\seealso{
Other example_data: 
\code{\link{example_biomarker_map_numeric}},
\code{\link{example_biomarker_map}},
\code{\link{example_biomarker_states_wide}},
\code{\link{example_biomarker_states}},
\code{\link{example_demography}},
\code{\link{example_exposure_force}},
\code{\link{example_exposure_probabilities}},
\code{\link{example_immune_histories_wide}},
\code{\link{example_immune_histories}},
\code{\link{example_model_pars_biphasic}},
\code{\link{example_model_pars_numeric}},
\code{\link{example_model_pars_typhoid}},
\code{\link{example_model_pars}},
\code{\link{example_observed_biomarker_states}}
}
\concept{example_data}
\keyword{datasets}
