% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_age_mod}
\alias{exposure_model_age_mod}
\title{Exposure Model Modified By Age}
\usage{
exposure_model_age_mod(i, t, e, g, lambdas, demography, age_mod, ...)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{g}{group}

\item{lambdas}{Force of infection array}

\item{demography}{Demography information}

\item{age_mod}{A tibble specifying the age modifier(how much each age affects probability of exposure); column names are age and modifier. All ages in simulation must have a modifier value within this tibble.}

\item{...}{}
}
\value{
A probability of exposure is returned
}
\description{
Probability of exposure depends on the force of infection at the current time t for group g modulated by the individual's age specified within age_mod
}
