% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_dem_age_mod}
\alias{exposure_model_dem_age_mod}
\title{Exposure Model Modified By Relevant Demographic Elements and Age}
\usage{
exposure_model_dem_age_mod(
  i,
  t,
  e,
  g,
  foe_pars,
  demography,
  mod,
  age_mod,
  t_in_year = 1,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{g}{group}

\item{foe_pars}{A 3D array providing the force of exposure for each exposure ID, group and time.}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{mod}{A tibble specifying the modifier(how much each input affects probability of exposure) for each demographic elements; column names are column, value, modifier. Entries in column and value must match format in demography table. All column and value combinations in demography must have a modifier value within this tibble.}

\item{age_mod}{A tibble specifying the age modifier(how much each age affects probability of exposure); column names are age and modifier. All ages in simulation must have a modifier value within this tibble.}

\item{t_in_year}{The number of time steps in a year; defaults to 1}

\item{...}{}
}
\value{
A probability of exposure is returned
}
\description{
Probability of exposure depends on the force of exposure at the current time t for group g modulated by relevant demographic elements (mod) and age (age_mod). This model combines exposure_model_V2 and exposure_model_V3
}
