% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_dem_mod}
\alias{exposure_model_dem_mod}
\title{Exposure Model Modified By Relevant Demographic Elements}
\usage{
exposure_model_dem_mod(i, t, e, g, lambdas, demography, mod, ...)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{g}{group}

\item{lambdas}{Force of infection array}

\item{demography}{Demography information}

\item{mod}{A tibble specifying the modifier(how much each input affects probability of exposure) for each demographic elements; column names are column, value, modifier. Entries in column and value must match format in demography table. All column and value combinations in demography must have a modifier value within this tibble.}

\item{...}{}
}
\value{
A probability of exposure is returned
}
\description{
Probability of exposure depends on the force of infection at the current time t for group g modulated by relevant demographic elements specified within the mod. Within mod, users can select which demographic elements affect the probability of exposure and by how much.
}
