% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_simple_FOE}
\alias{exposure_model_simple_FOE}
\title{Exposure Model Simple- Force of Exposure (FOE)}
\usage{
exposure_model_simple_FOE(i, t, x, g, foe_pars, demography, ...)
}
\arguments{
\item{i}{individual}

\item{t}{time}

\item{x}{exposure}

\item{g}{group}

\item{foe_pars}{A 3D array providing the force of exposure for each exposure ID, group and time.}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{...}{Additional arguments}
}
\value{
A probability of exposure is returned
}
\description{
This is a simple exposure model where the probability of exposure depends on the force of exposure at that time for that group
}
\examples{
times <- seq(1,365,by=1)
## Create fixed FOI (force of infection) for one exposure type for two groups
n_groups <- 2
n_exposures <- 1
foe_pars <- array(NA, dim=c(n_groups,length(times),n_exposures))
foe_pars[1,,] <- 0.01
foe_pars[2,,] <- 0.005
## Solve the model for each time point for each group. Note that the i argument is not used, 
## but is kept for compatibility with other functions.
foe <- matrix(NA, nrow=n_groups,ncol=length(times))
for(g in 1:nrow(foe)){
    foe[g,] <- unlist(sapply(times, function(t) exposure_model_simple_FOE(NULL, t, 1, g, foe_pars, NULL)))
}
}
