% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_models.R
\name{exposure_model_sir}
\alias{exposure_model_sir}
\title{SIR Exposure Model}
\usage{
exposure_model_sir(i, t, x, g, foe_pars, demography = NULL, time_res = 1, ...)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{g}{group}

\item{foe_pars}{Data frame containing SIR model parameters for each group and exposure combination. Variable names: x (exposure ID), g (group ID), name (parameter name), value (parameter value). Parameters needed are: beta (transmission rate), gamma (recovery rate), I0 (per capita infected population seed size), R0 (per capita recovered population seed size) and t0 (seeding time).}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{time_res}{Time steps to solve the ODEs. Set lower for higher accuracy.}

\item{...}{}
}
\value{
Probability of exposure for the requested time step
}
\description{
SIR Exposure Model
}
