% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_demography.R
\name{generate_pop_demography}
\alias{generate_pop_demography}
\title{Generate A Population Demography Data Set}
\usage{
generate_pop_demography(
  N,
  times,
  birth_times = NULL,
  limit = 0,
  removal_min,
  removal_max,
  prob_removal,
  aux = NULL
)
}
\arguments{
\item{N}{The number of individuals in the simulation}

\item{times}{The total number of time steps in the simulation}

\item{birth_times}{A vector of birth times for each individual; defaults to NULL; if birth_times is not specified then the function will simulate birth times for each individual}

\item{limit}{This number limits the last month an individual is born; if you want to ensure that all individuals simulated are above age of vaccination then enter the time step at which individuals are eligible for vaccination; defaults to 0 which allows individuals to be born up until the second to last step which ensures that each individual is alive for sampling at the last time step; defaults to 0}

\item{removal_min}{The minimum age at which an individual can be removed from the population}

\item{removal_max}{The maximum age at which an individual can be removed from the population}

\item{prob_removal}{The probability that an individual will be removed from the population}

\item{aux}{A list of the demography columns, the variable options and their distributions; defaults to NULL}
}
\value{

}
\description{
Generate A Population Demography Data Set
}
\examples{
generate_pop_demography(10, 1:120, limit=0, removal_min=0, removal_max=120, prob_removal=0.3)

aux <- list("Sex"=list("name"="sex","options"=c("male", "female"), "distribution"=c(0.5,0.5)),"Group"=list("name"="group","options"=c("1", "2", "3", "4"), "distribution"=c(0.25,0.25,0.25,0.25)) )
generate_pop_demography(10, 1:120, limit=0, removal_min=0, removal_max=120, prob_removal=0.3, aux=aux)
}
