% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_all_successful}
\alias{immunity_model_all_successful}
\title{Immunity Model Where All Exposures Are Successful}
\usage{
immunity_model_all_successful(
  i,
  t,
  e,
  exposure_histories,
  antibody_states,
  demography,
  antigen_map,
  theta,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{antibody_states}{True antibody titers for all individuals across all time steps and antigens}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{antigen_map}{A table specifying the relationship between exposure IDs and antigen IDs}

\item{theta}{A tibble of parameters needed for the immunity model}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
Probability of success is 1 so all exposure events are successful
}
