% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_vacc_ifxn_titer_prot}
\alias{immunity_model_vacc_ifxn_titer_prot}
\title{Immunity Model For Vaccination Events and Natural Infection Events With Titer-Mediated Protection}
\usage{
immunity_model_vacc_ifxn_titer_prot(
  i,
  t,
  x,
  exposure_histories,
  antibody_states,
  demography,
  biomarker_map,
  model_pars,
  max_events,
  vacc_exposures,
  vacc_age = 1,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{x}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure events}

\item{antibody_states}{True antibody titers for all individuals across all time steps and biomarkers}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{biomarker_map}{A table specifying the relationship between exposure events and biomarkers}

\item{model_pars}{A tibble of parameters needed for the immunity model}

\item{max_events}{A vector of the maximum number of successful exposure events possible for each exposure event; If an exposure type is not a vaccination event then input NA}

\item{vacc_exposures}{A vector of exposure IDs (x) which represents vaccination events}

\item{vacc_age}{A vector of the minimum age at which an individual is eligible for vaccination for each exposure event; If an exposure event is not a vaccination event then input NA}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
This immunity model should be used if exposures represent vaccination and natural infection events. The probability of a successful vaccination exposure event depends on the number of vaccines received prior to time t while the probability of successful infection is dependent on the titer at the time of exposure and the total number of successful infections prior to that point.
}
