% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_vacc_only}
\alias{immunity_model_vacc_only}
\title{Immunity Model For Vaccination Events Only}
\usage{
immunity_model_vacc_only(
  i,
  t,
  e,
  exposure_histories,
  antibody_states,
  demography,
  biomarker_map,
  model_pars,
  max_vacc_events,
  vacc_age,
  ...
)
}
\arguments{
\item{i}{Individual}

\item{t}{time}

\item{e}{exposure}

\item{exposure_histories}{An array of exposure histories across all individuals, time steps and exposure IDs}

\item{antibody_states}{True antibody titers for all individuals across all time steps and biomarkers}

\item{demography}{A tibble of relevant demographic information for each individual in the simulation.}

\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{model_pars}{A tibble of parameters needed for the immunity model}

\item{max_vacc_events}{A vector of the maximum number of vaccination events possible for each exposure type; If an exposure type is not a vaccination event then input NA}

\item{vacc_age}{A vector of the minimum age at which an individual is eligible for vaccination for each exposure type; If an exposure type is not a vaccination event then input NA}

\item{...}{}
}
\value{
A probability of successful exposure is returned
}
\description{
This immunity model should only be used if all exposures are vaccination events. The probability of successful exposure(vaccination event) depends on the number of vaccines an individual has received prior to time t. If the individual is under the maximum vaccinations allotted then the probability of successful exposure is 1.
}
