% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunity_models.R
\name{immunity_model_vacc_only}
\alias{immunity_model_vacc_only}
\title{Vaccine-mediated immunity model}
\usage{
immunity_model_vacc_only(
  i,
  t,
  x,
  exposure_histories,
  biomarker_states,
  demography,
  biomarker_map,
  model_pars,
  max_vacc_events,
  vacc_age,
  ...
)
}
\arguments{
\item{i}{integer for the individual ID}

\item{t}{integer for the time period}

\item{x}{integer for the exposure ID}

\item{exposure_histories}{a 3D array of exposure histories for all individuals, time steps and exposure IDs}

\item{biomarker_states}{an 3D array of biomarker states (biomarker quantities) for all individuals, time steps and biomarker IDs}

\item{demography}{a tibble of demographic information for each individual in the simulation}

\item{biomarker_map}{a table specifying the relationship between exposure IDs and biomarker IDs}

\item{model_pars}{a tibble of parameters needed for the immunity model}

\item{max_vacc_events}{a vector of the maximum number of vaccination events possible for each exposure type; If an exposure type is not a vaccination event then input \code{NA}}

\item{vacc_age}{a vector of the minimum age at which an individual is eligible for vaccination for each exposure event; If an exposure event is not a vaccination event then input \code{NA}}

\item{...}{}
}
\value{
The probability of successful exposure
}
\description{
This immunity model should only be used if all exposures are vaccination events. The probability of successful exposure (vaccination event) depends on the number of vaccines an individual has received prior to time t. If the individual is under the maximum vaccinations allotted then the probability of successful exposure event is 1.
}
\examples{
tmp_exposure_history <- array(0, dim=c(1, 10, 1))
tmp_exposure_history[1,1,1] <- 1
tmp_demography <- tibble(i=1, birth=1)
immunity_model_vacc_only(1,8,1,tmp_exposure_history,NULL, tmp_demography,NULL, NULL,max_vacc_events=2,vacc_age=5)
}
