% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_funcs.R
\name{multiple_strains}
\alias{multiple_strains}
\title{Boost/wane model}
\usage{
multiple_strains(tis, y0, params, times)
}
\arguments{
\item{tis}{vector of times of infection eg. c(0, 45, 70)}

\item{y0}{vector of baseline HAI values for each infecting strain}

\item{params}{list of parameters for the model. This should be params[[1]] = mu_pars (matrix); params[[2]] = tp_pars (matrix); params[[3]] = m_pars (matrix); params[[4]] = STOCHASTIC (bool).}

\item{times}{vector of times in days.}
}
\value{
matrix of times and longitundinal simulation for each infecting strain
}
\description{
Data generating function for a multiple strain, boosting and waning serological model.
}
\seealso{
\code{\link{single_strain}}
}

