% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_to_lognormal_sd.R
\name{normal_to_lognormal_sd}
\alias{normal_to_lognormal_sd}
\title{Convert Normal Distribution Standard Deviation to Log-Normal Distribution Standard Deviation}
\usage{
normal_to_lognormal_sd(normmean, normsd)
}
\arguments{
\item{normmean}{The normal distribution mean}

\item{normsd}{The normal distribution standard deviation}
}
\value{
The log-normal distribution standard deviation is printed
}
\description{
Convert Normal Distribution Standard Deviation to Log-Normal Distribution Standard Deviation
}
\examples{
normal_to_lognormal_sd(5000,400)
}
