% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{observation_model_continuous}
\alias{observation_model_continuous}
\title{Observation Model For Continuous Assays With No Added Noise}
\usage{
observation_model_continuous(biomarker_states, model_pars, ...)
}
\arguments{
\item{biomarker_states}{True biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of observation model parameters}

\item{...}{}
}
\value{
biomarker_states is returned with a new column for observed biomarker quantities
}
\description{
This observation model observes the latent biomarker quantities given a continuous assay with no added noise. Therefore the true biomarker quantity is the observed biomarker quantity.
}
