% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{observation_model_continuous_noise}
\alias{observation_model_continuous_noise}
\title{Observation Model For Continuous Assays With Added Noise}
\usage{
observation_model_continuous_noise(
  biomarker_states,
  model_pars,
  sensitivity = 1,
  specificity = 1,
  ...
)
}
\arguments{
\item{biomarker_states}{True biomarker quantities for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of observation model parameters}

\item{sensitivity}{number between 0 and 1 to describe the assay's sensitivity; defaults to 1}

\item{specificity}{number between 0 and 1 to describe the assay's specificity; defaults to 1}

\item{...}{}
}
\value{
biomarker_states is returned with a new column for observed biomarker quantities
}
\description{
This observation model observes the latent biomarker quantities given a continuous assay with added noise. The added noise represents assay variability and is done by sampling from a distribution with the latent biomarker quantity as the mean and the measurement error as the standard deviation. The observation standard deviation and distribution are defined within model_pars as the “obs_sd” parameter. The user can also use the optional sensitivity and specificity arguments to account for assay sensitivity and specificity.
}
