% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{observation_model_discrete}
\alias{observation_model_discrete}
\title{Observation model for discrete assays with no added noise}
\usage{
observation_model_discrete(biomarker_states, model_pars, cutoffs, ...)
}
\arguments{
\item{biomarker_states}{tibble containing true biomarker quantities for all individuals across all time steps and biomarkers. Variables should include: 1) i: the individual ID; 2) t: the time period; 3) b: the biomarker ID; 4) value: the latent biomarker quantity for the given i, t and b}

\item{model_pars}{a tibble containing information for all parameters needed to simulate the observation process. This should usually contain: 1) exposure_id: numeric exposure ID; 2) biomarker_id: numeric biomarker ID; 3) name: the character name of the parameter; 4) mean: numeric mean of this parameter distribution; 5) sd: the numeric standard deviation of the parameter distribution}

\item{cutoffs}{a matrix containing the assay cutoffs for each biomarker. Each row contains all of the cutoffs for that biomarker starting with 0. For example, all of the cutoffs for the assay measuring biomarker 1 quantity are in the first row of this matrix.}

\item{...}{Additional arguments}
}
\value{
\code{biomarker_states} is returned with a new column, \code{observed}, for observed biomarker quantities
}
\description{
This observation model observes the latent biomarker quantities given a discrete assay with user-specified ranges within the "discrete" argument and no added noise.
}
\examples{
breaks <- seq(0,8,by=1)
cutoffs <- matrix(breaks,nrow=1,ncol=length(breaks))
observation_model_discrete(example_biomarker_states, NULL,cutoffs)
}
\seealso{
Other observation_model: 
\code{\link{observation_model_continuous_bounded_noise}()},
\code{\link{observation_model_continuous_bounded}()},
\code{\link{observation_model_continuous_noise}()},
\code{\link{observation_model_continuous}()},
\code{\link{observation_model_discrete_noise}()}
}
\concept{observation_model}
