% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_models.R
\name{observation_model_discrete_noise}
\alias{observation_model_discrete_noise}
\title{Observation Model For Discrete Assays With Added Noise}
\usage{
observation_model_discrete_noise(antibody_states, model_pars, cutoffs, ...)
}
\arguments{
\item{antibody_states}{True antibody titers for all individuals across all time steps and biomarkers}

\item{model_pars}{Tibble of observation model parameters}

\item{cutoffs}{A matrix containing the assay cutoffs for each biomarker. Each row contains all of the cutoffs for that biomarker starting with 0. For example, all of the cutoffs for the assay measuring biomarker 1 specific antibodies are in the first row of this matrix.}

\item{...}{}
}
\value{
antibody_states is returned with a new column for observed titers
}
\description{
This observation model observes the latent titer values given a discrete assay with user-specified ranges within discrete and added noise. The added noise represents assay variability and is done by sampling from a distribution with the latent antibody titer as the mean and the measurement error as the standard deviation. The observation standard deviation and distribution are defined within model_pars as the “obs_sd” parameter.
}
