% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_antibody_model}
\alias{plot_antibody_model}
\title{Plots example trajectories of the provided antibody kinetics model}
\usage{
plot_antibody_model(
  antibody_model,
  N = 100,
  times = seq(1, 50, by = 1),
  model_pars,
  biomarker_map,
  demography = NULL,
  draw_parameters_fn = draw_parameters_fixed_fx,
  ...
)
}
\description{
Plots example trajectories of the provided antibody kinetics model
}
