% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_antibody_model}
\alias{plot_antibody_model}
\title{Plots example trajectories of the provided antibody kinetics model}
\usage{
plot_antibody_model(
  antibody_model,
  N = 100,
  times = seq(1, 50, by = 1),
  model_pars,
  biomarker_map,
  demography = NULL,
  draw_parameters_fn = draw_parameters_fixed_fx,
  ...
)
}
\arguments{
\item{antibody_model}{A function determining the latent biomarker kinetics as a function of immune history and within-host kinetics parameters (\code{model_pars})}

\item{N}{Number of trajectories to simulate}

\item{times}{Vector of times to solve model over}

\item{model_pars}{A tibble of parameters needed for the antibody kinetics model, immunity model, observation model and the draw_parameters function}

\item{biomarker_map}{A tibble specifying the relationship between exposure IDs and biomarker IDs with two variables: 1) exposure_id: the numeric exposure ID; 2) biomarker_id: the numeric biomarker ID}

\item{demography}{A tibble of demographic information for each individual in the simulation. At a minimum this tibble requires 1 column (i) where all individuals in the simulation are listed by row. This is used to calculate the sample population size. Additional variables can be added by the user, e.g., birth and removal times, see \code{\link{generate_pop_demography}} If not specified, the model will assume that birth time is the initial time point and removal time is the final time point across all individuals}

\item{draw_parameters_fn}{Pointer to function used for drawing random kinetics parameters, see \code{\link{draw_parameters_fixed_fx}}}

\item{...}{Any additional arguments needed for the models}
}
\value{
A ggplot2 object
}
\description{
Plots example trajectories of the provided antibody kinetics model
}
\examples{
library(dplyr)
model_pars <- reformat_biomarker_map(example_model_pars_biphasic)\%>\% tidyr::drop_na()
draw_parameters_random_fx(1,1,1,1,NULL,NULL,model_pars)
biomarker_map <- dplyr::tibble(exposure_id=c(1,1,2),biomarker_id=c(1,2,1))
plot_antibody_model(antibody_model_biphasic, 50, model_pars=model_pars,
draw_parameters_fn = draw_parameters_random_fx, biomarker_map=biomarker_map)
plot_antibody_model(antibody_model_biphasic, 50, model_pars=model_pars,
draw_parameters_fn = draw_parameters_fixed_fx, biomarker_map=biomarker_map)
}
