% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_exposure_force}
\alias{plot_exposure_force}
\title{Plot individual exposure probabilities across time for all individuals and exposure events}
\usage{
plot_exposure_force(exposure_force_long)
}
\arguments{
\item{exposure_force_long}{The reshaped data set containing the probability of an exposure event for individuals at all time steps for each exposure event}
}
\value{
A plot of individual exposure probabilities across time for all exposure events is returned
}
\description{
Plot individual exposure probabilities across time for all individuals and exposure events
}
\examples{
plot_exposure_force(example_exposure_force)
}
