% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_exposure_force}
\alias{plot_exposure_force}
\title{Plot Force of Exposure for Each Individual Across Time For All Individuals And Exposure IDs}
\usage{
plot_exposure_force(exposure_force_long)
}
\arguments{
\item{exposure_force_long}{The reshaped data set containing exposure probability for individuals at all time steps for each exposure ID}
}
\value{
A plot of force of exposure for across time for all individuals and exposure IDs is returned
}
\description{
Plot Force of Exposure for Each Individual Across Time For All Individuals And Exposure IDs
}
\examples{
plot_exposure_force(example_force_long)
}
