% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_exposure_histories}
\alias{plot_exposure_histories}
\title{Plot Individual Exposure Histories}
\usage{
plot_exposure_histories(exposure_histories)
}
\arguments{
\item{expsoure_histories}{The reshaped data set containing exposure history for individuals at all time steps for each exposure event}
}
\value{
A plot of individual exposures histories across time for all individuals and exposure events is returned
}
\description{
Plot Individual Exposure Histories
}
\examples{
plot_exposure_histories(example_exposure_histories)
}
