% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_exposure_model}
\alias{plot_exposure_model}
\title{Plots the probability of exposure over time for the provided exposure models}
\usage{
plot_exposure_model(
  exposure_model,
  times,
  n_groups = 1,
  n_exposures = 1,
  foe_pars,
  demography = NULL,
  ...
)
}
\arguments{
\item{exposure_model}{A function calculating the probability of exposure given the foe_pars array}

\item{times}{Vector of times to solve model over}

\item{n_groups}{Number of groups corresponding to \code{foe_pars}}

\item{n_exposures}{Number of exposure types corresponding to \code{foe_pars}}

\item{foe_pars}{Generic object containing all parameters needed to solve \code{exposure_model}}

\item{demography}{A tibble of necessary demographic information for each individual in the simulation. At a minimum this tibble requires 1 column (i) where all individuals in the simulation are listed by row. This is used to calculate the sample population size. Additional variables can be added by the user, e.g., birth and removal times. If not specified, the model will assume that birth time is the initial time point and removal time is the final time point across all individuals.}
}
\value{
A ggplot2 object
}
\description{
Plots the probability of exposure over time for the provided exposure models
}
\examples{
## Basic exposure model with demography modifier
foe_pars <- array(NA, dim=c(n_groups,length(times),n_exposures))
foe_pars[1,,1] <- 0.01
foe_pars[1,,2] <- 0.005 
demography <- tibble(i = rep(1:n_indiv, each=n_times), t=rep(times,2),
                     SES=rep(c("low","high"),each=n_times))
dem_mod <- tibble(exposure_id=c(1,1,2,2),column=c("SES","SES","SES","SES"),
                 value=c("low","high","low","high"),modifier=c(1,0.75,1,0.5))

plot_exposure_model(exposure_model=exposure_model_dem_mod, 1:365,
                    1,2,foe_pars,demography = demography,dem_mod=dem_mod)
                    
## SIR model with two groups and two exposure types                 
foe_pars <- bind_rows(
                      tibble(x=1,g=1,name=c("beta","gamma","I0","R0","t0"),value=c(0.3,0.2,0.00001,0,0)),
                      tibble(x=2,g=1,name=c("beta","gamma","I0","R0","t0"),value=c(0.35,0.25,0.00001,0,200)),
                      tibble(x=1,g=2,name=c("beta","gamma","I0","R0","t0"),value=c(0.5,0.2,0.00005,0,0)),
                      tibble(x=2,g=2,name=c("beta","gamma","I0","R0","t0"),value=c(0.27,0.2,0.00001,0,50))
                      )
plot_exposure_model(exposure_model_sir, seq(1,365,by=1),n_groups = 2,n_exposures = 2,foe_pars=foe_pars)
                    
}
