% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{plot_subset_individuals_history}
\alias{plot_subset_individuals_history}
\title{Plot Antibody States and Exposure Histories For A Subset Of Individuals}
\usage{
plot_subset_individuals_history(
  titers,
  exposure_histories,
  subset,
  demography,
  removal = FALSE
)
}
\arguments{
\item{titers}{The reshaped data set containing antibody titer for individuals at all time steps for each biomarker}

\item{exposure_histories}{The reshaped data set containing exposure history for individuals at all time steps for each exposure evnt}

\item{subset}{The number of individuals you want to plot}

\item{demography}{Tibble of removal time for each individual}

\item{removal}{Set to TRUE if individuals are removed during the simulation and removal time is present in demogrpahy; defaults to FALSE}
}
\value{

}
\description{
Plot Antibody States and Exposure Histories For A Subset Of Individuals
}
\examples{
plot_subset_individuals_history(example_antibody_states,example_exposure_histories,3,example_demography)
}
