% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{precomputation_checks}
\alias{precomputation_checks}
\title{Check exposure model for pre-computation possibilities}
\usage{
precomputation_checks(
  N,
  times,
  exposure_ids,
  groups,
  exposure_model,
  foe_pars,
  demography,
  VERBOSE,
  check_correct = FALSE,
  ...
)
}
\description{
Checks if the provided exposure model can be solved in advance of the main simulation in fewer function calls than would be expected in the main simulation. Note that there is some overhead to checking for pre-computation, so this may not be quicker and may actually slow down the simulation. However, in some cases where many individuals share identical exposure probabilities, running this function can lead to massive speed ups.
}
