% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{reformat_model_pars}
\alias{reformat_model_pars}
\title{Reformat model_pars to numeric exposure and biomarker format needed for runserosim}
\usage{
reformat_model_pars(biomarker_map, model_pars)
}
\arguments{
\item{biomarker_map}{A table specifying the relationship between exposure IDs and biomarker IDs}

\item{model_pars}{A tibble of parameters needed for the antibody kinetics model, immunity model, observation model and the draw_parameters function}
}
\value{
biomarker_map is returned with unique numeric inputs for exposure and biomarker IDs
}
\description{
This function will reformat the model_pars argument so that exposure_ID and biomarker_ID are both numeric
}
