% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{simulate_removal_times}
\alias{simulate_removal_times}
\title{Simulate removal times for individuals}
\usage{
simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min = 0,
  removal_max = max(times),
  prob_removal = 0
)
}
\arguments{
\item{N}{The number of individuals in the simulation}

\item{times}{A vector of each time step in the simulation}

\item{birth_times}{A vector of birth times for each individual; defaults to NULL; if \code{birth_times} is not specified then the function will simulate uniformly distributed birth times for each individual from the times vector}

\item{removal_min}{The minimum age at which an individual can be removed from the population. Defaults to 0}

\item{removal_max}{The maximum age at which an individual can be removed from the population. Defaults to \code{max(times)}}

\item{prob_removal}{The probability that an individual will be removed from the population during the simulation, representing e.g., death or study attrition. If set to \code{NA}, then removal time will be \code{max(times)+1}}
}
\value{
A vector of all individual's removal times is returned. \code{NA} represents no removal
}
\description{
This function simulates uniformly distributed removal times for a specified number of individuals from the provided times vector. This might represent e.g., death or study attrition.
}
\examples{
## Simulate random removal times for all individuals; Individuals have a 0.4 probability of being removed at some time after they are 10 time steps old and before they are 99 time steps old 
birth_times<-simulate_birth_times(500, 1:100, age_min=9) 
simulate_removal_times(500,1:100,birth_times, removal_min=10,removal_max=99, prob_removal=0.4)
}
\seealso{
Other demography: 
\code{\link{generate_pop_demography}()},
\code{\link{simulate_birth_times}()}
}
\concept{demography}
