% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pop_demography.R
\name{simulate_removal_times}
\alias{simulate_removal_times}
\title{Simulate  Removal Times for Individuals}
\usage{
simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min,
  removal_max,
  prob_removal
)
}
\arguments{
\item{N}{The number of individuals in the simulation}

\item{times}{The total number of time steps in the simulation}

\item{birth_times}{A vector of all individual's birth times}

\item{removal_min}{The minimum age at which an individual can be removed from the population}

\item{removal_max}{The maximum age at which an individual can be removed from the population}

\item{prob_removal}{The probability that an individual will be removed from the population}
}
\value{
A vector of all individual's removal times is returned
}
\description{
Simulate  Removal Times for Individuals
}
